+++ src/main/kotlin/com/wafflestudio/account/api/client/KakaoClient.kt
@@ -1,5 +1,6 @@
 package com.wafflestudio.account.api.client
 
+import com.fasterxml.jackson.annotation.JsonProperty
 import com.wafflestudio.account.api.domain.account.enum.SocialProvider
 import kotlinx.coroutines.reactor.awaitSingleOrNull
 import org.springframework.http.MediaType
@@ -72,3 +73,17 @@ class KakaoClient(
         }
     }
 }
+
+data class KakaoOAuth2TokenResponse(
+    @JsonProperty("token_type")
+    val tokenType: String?,
+    @JsonProperty("access_token")
+    val accessToken: String,
+    @JsonProperty("expires_in")
+    val expiresIn: Int?,
+    @JsonProperty("refresh_token")
+    val refreshToken: String?,
+    @JsonProperty("refresh_token_expires_in")
+    val refreshTokenExpiresIn: Int?,
+    val scope: String?,
+)
